requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
cryptography==41.0.8
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0

